Improvements on workspace file loading
